Console Schema Finished.
1. Got rid of habtm related tables from being automatically included in console generated schema files.

2. Got rid of aros, acos, and aros_acos from being automatically included in console generated schema files.

3. Add the header necessary to work with the zuha upgrade functions

This is all good, because now you can upgrade the zuha database without worrying about which version the settings state, because we compare the database to the schema files and see what the difference is and update accordingly.

Also its really fast to generate a schema file after making a db update. Just run the console command...

cake zuha_schema generate --plugin SomePluginName (where SomePluginName equals the actual plugin name, or just omit the --plugin option to generate the core app schema.)

IMPORTANT NOTE : If you rename a column it does take a manual update to the schema file to add the table name, and old and new names of the column.  But if you just add or drop it will be all set.  The manual update allows us to automatically copy data from the old table column to the new table column so that no data is lost.